Completed refactoring io utils, no code changes
