+++ mlcolvar/utils/io/graphs/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ["create_dataset_from_trajectories",
+           "create_pdb_from_xyz"]
+
+# not imported by default as they depend on optional libraries (pandas, scikit-learn or KDEpy)
+from .common import *
+from .mdtraj import *
+from .ase import *
