+++ mlcolvar/utils/io/graphs/common.py
@@ -1,59 +1,19 @@
-"""Input/output functions."""
-try:
-    import pandas as pd
-except ImportError as e:
-    raise ImportError(
-        "pandas is required to use the i/o utils (mlcolvar.utils.io)\n", e
-    )
-
 import numpy as np
 import torch
 import os
 from typing import Union, List, Tuple
 import mdtraj
-from warnings import warn
-
 
 from mlcolvar.utils.io._utils import _download_temp_file
+from mlcolvar.utils.io.graphs._utils import *
+from mlcolvar.utils.io.graphs.mdtraj import *
+from mlcolvar.utils.io.graphs.ase import *
+
 from mlcolvar.data import DictDataset
-from mlcolvar.data.graph.atomic import AtomicNumberTable, Configuration, Configurations
-from mlcolvar.data.graph.utils import create_dataset_from_configurations
+from mlcolvar.data.graph.atomic import AtomicNumberTable
 
-__all__ = ["create_dataset_from_configurations", 
-           "create_dataset_from_trajectories",
-           "create_pdb_from_xyz"]
+__all__ = ["create_dataset_from_trajectories"]
 
-
-def create_pdb_from_xyz(input_filename: str, output_filename: str) -> str:
-    """
-    Convert the first frame of an XYZ file into a PDB file using ASE.
-    This pdb file can then serve as the topology for MDTraj.
-
-    Parameters:
-        input_filename: Path to the input .xyz file.
-        output_filename: Path to the output .pdb file.
-
-    Returns:
-        The path to the generated PDB file.
-    """
-    # Import ASE 
-    try:
-        from ase.io import read, write
-        from ase import Atoms
-    except ImportError as e:
-        raise ImportError("ASE is required for xyz to pdb conversion.", e)
-
-    atoms: Atoms = read(input_filename, index=0)
-
-    if (atoms.cell == 0).all():
-        warn("A topology file was generated from the xyz trajectory file but no cell information were provided!")
-    if not atoms.pbc.any():
-        warn("A topology file was generated from the xyz trajectory file but no PBC information were provided!")
-    elif not atoms.pbc.all():
-        warn( f"Partial PBC are not supported! The provided input has pbc {atoms.pbc}")
-
-    write(output_filename, atoms, format='proteindatabank')
-    return output_filename
 
 def create_dataset_from_trajectories(
     trajectories: Union[List[str], str],
@@ -326,410 +286,6 @@ def create_dataset_from_trajectories(
         return dataset, trajectories_in_memory
     else:
         return dataset
-
-def load_traj_with_mdtraj(trajectory: str, 
-                          topology: str, 
-                          selection: str):
-    # load trajectory
-        traj = mdtraj.load(trajectory, top=topology)
-        traj.top = mdtraj.core.trajectory.load_topology(topology)
-        
-        # mdtraj does not load cell info from xyz, so we use ASE and add it
-        _, ext = os.path.splitext(trajectory)
-        if ext.lower() == ".xyz":
-            try:
-                from ase.io import read
-            except ImportError as e:
-                raise ImportError("ASE is required for creating the graph from a .xyz file.", e)
-            ase_atoms = read(trajectory, index=':')
-            ase_cells = np.array([a.get_cell().array for a in ase_atoms], dtype=float)
-            # the pdb for the topology are in nm, ase work in A so we need to scale it
-            traj.unitcell_vectors = ase_cells/10
-
-        if selection is not None:
-            subset = traj.top.select(selection)
-            assert len(subset) > 0, (
-                'No atoms will be selected with selection string '
-                + '"{:s}"!'.format(selection)
-            )
-            traj = traj.atom_slice(subset)
-        
-        return traj
-
-def _as_torch_if_array(x):
-    if isinstance(x, torch.Tensor):
-        return x.detach().cpu()
-    if isinstance(x, np.ndarray):
-        return torch.as_tensor(x)
-    return x
-
-
-def _to_torch_tensor(x, dtype=torch.get_default_dtype()):
-    if isinstance(x, torch.Tensor):
-        return x.detach().cpu().to(dtype=dtype)
-    return torch.as_tensor(x, dtype=dtype)
-
-
-def _get_selected_frame_indices(n_frames: int, load_arg: dict = None) -> List[int]:
-    if load_arg is None:
-        return list(range(0, n_frames, 1))
-    start = load_arg.get('start', 0)
-    stop = load_arg.get('stop', None)
-    stride = load_arg.get('stride', 1)
-    if stop is None:
-        stop = n_frames
-    return list(range(start, stop, stride))
-
-
-def _normalize_trajectory_labels(trajectory_labels, n_traj: int):
-    if trajectory_labels is None:
-        return [torch.tensor([[i]], dtype=torch.get_default_dtype()) for i in range(n_traj)]
-
-    trajectory_labels = _as_torch_if_array(trajectory_labels)
-
-    if isinstance(trajectory_labels, torch.Tensor):
-        if trajectory_labels.ndim == 0:
-            trajectory_labels = trajectory_labels.reshape(1).repeat(n_traj)
-            trajectory_labels = trajectory_labels.tolist()
-        elif trajectory_labels.ndim >= 1 and trajectory_labels.shape[0] == n_traj:
-            trajectory_labels = [trajectory_labels[i] for i in range(n_traj)]
-        else:
-            if n_traj == 1:
-                raise ValueError(
-                    f"trajectory_labels has length {trajectory_labels.shape[0]} for a single trajectory. Use graph_labels for per-frame targets."
-                )
-            raise ValueError(
-                f"trajectory_labels first dimension ({trajectory_labels.shape[0]}) must match number of trajectories ({n_traj})."
-            )
-
-    if not isinstance(trajectory_labels, (list, tuple)):
-        trajectory_labels = [trajectory_labels]
-
-    if len(trajectory_labels) != n_traj:
-        raise ValueError(
-            f"Number of trajectory labels ({len(trajectory_labels)}) must match number of trajectories ({n_traj})."
-        )
-
-    normalized = []
-    for item in trajectory_labels:
-        item = _as_torch_if_array(item)
-        if isinstance(item, torch.Tensor):
-            if item.ndim == 0:
-                normalized.append(item.reshape(1, 1).to(dtype=torch.get_default_dtype()))
-            else:
-                normalized.append(item.reshape(-1, 1).to(dtype=torch.get_default_dtype()))
-        elif np.isscalar(item):
-            normalized.append(torch.tensor([[item]], dtype=torch.get_default_dtype()))
-        else:
-            arr = _to_torch_tensor(item)
-            if arr.ndim == 0:
-                normalized.append(arr.reshape(1, 1))
-            else:
-                normalized.append(arr.reshape(-1, 1))
-    return normalized
-
-
-def _broadcast_trajectory_to_graph_labels(trajectory_labels, frame_counts: List[int]):
-    trajectory_labels = _normalize_trajectory_labels(trajectory_labels, len(frame_counts))
-    broadcast = []
-    for i, n_frames in enumerate(frame_counts):
-        y = trajectory_labels[i].reshape(1, -1)
-        broadcast.append(y.repeat(n_frames, 1))
-    return broadcast
-
-
-
-
-def _normalize_frame_level_labels(labels, frame_counts: List[int], name: str):
-    n_traj = len(frame_counts)
-
-    if labels is None:
-        return [None for _ in range(n_traj)]
-
-    labels = _as_torch_if_array(labels)
-
-    if n_traj == 1:
-        n_frames = frame_counts[0]
-        if isinstance(labels, torch.Tensor) and labels.ndim >= 1 and labels.shape[0] == n_frames:
-            labels = [labels]
-        elif (
-            isinstance(labels, (list, tuple))
-            and len(labels) == n_frames
-            and len(labels) != n_traj
-        ):
-            warn(
-                f"`{name}` was passed as a flat list/tuple of length {n_frames} for a single trajectory; "
-                "interpreting it as frame-level labels. To avoid ambiguity, prefer passing a tensor/array "
-                "or wrapping per-trajectory labels as [labels]."
-            )
-            labels = [_to_torch_tensor(labels)]
-
-    if not isinstance(labels, (list, tuple)):
-        raise ValueError(f"{name} must be a list/tuple with one entry per trajectory.")
-
-    if len(labels) != n_traj:
-        raise ValueError(
-            f"{name} has {len(labels)} entries but number of trajectories is {n_traj}."
-        )
-
-    normalized = []
-    for i, item in enumerate(labels):
-        n_frames = frame_counts[i]
-        if item is None:
-            normalized.append(None)
-            continue
-
-        item = _as_torch_if_array(item)
-        arr = _to_torch_tensor(item)
-
-        if arr.ndim == 0:
-            arr = arr.reshape(1, 1).repeat(n_frames, 1)
-        elif arr.ndim == 1:
-            if arr.shape[0] != n_frames:
-                raise ValueError(
-                    f"{name}[{i}] length ({arr.shape[0]}) must match selected frames ({n_frames})."
-                )
-            arr = arr.reshape(-1, 1)
-        elif arr.ndim == 2:
-            if arr.shape[0] != n_frames:
-                raise ValueError(
-                    f"{name}[{i}] first dimension ({arr.shape[0]}) must match selected frames ({n_frames})."
-                )
-        elif arr.ndim == 3 and name == 'node_labels':
-            if arr.shape[0] != n_frames:
-                raise ValueError(
-                    f"{name}[{i}] first dimension ({arr.shape[0]}) must match selected frames ({n_frames})."
-                )
-        else:
-            raise ValueError(f"Unsupported shape for {name}[{i}]: {tuple(arr.shape)}.")
-
-        normalized.append(arr)
-
-    return normalized
-
-
-def _normalize_graph_target_inputs(
-    trajectories: List[mdtraj.Trajectory],
-    load_args: list,
-    trajectory_labels=None,
-    graph_labels=None,
-    node_labels=None,
-):
-    n_traj = len(trajectories)
-    frame_counts = [
-        len(_get_selected_frame_indices(
-            n_frames=len(traj),
-            load_arg=load_args[i] if load_args is not None else None,
-        ))
-        for i, traj in enumerate(trajectories)
-    ]
-
-    if trajectory_labels is not None and graph_labels is not None:
-        raise ValueError("Only one of `trajectory_labels` or `graph_labels` can be provided.")
-
-    if trajectory_labels is None and graph_labels is None:
-        trajectory_labels = [i for i in range(n_traj)]
-
-    if graph_labels is None:
-        graph_labels = _broadcast_trajectory_to_graph_labels(
-            trajectory_labels=trajectory_labels,
-            frame_counts=frame_counts,
-        )
-    else:
-        graph_labels = _normalize_frame_level_labels(graph_labels, frame_counts, name='graph_labels')
-
-    node_labels = _normalize_frame_level_labels(node_labels, frame_counts, name='node_labels')
-
-    return graph_labels, node_labels
-
-def dataset_from_mdtraj_trajectories(trajectories: List[mdtraj.Trajectory],
-                                     graph_labels: List,
-                                     node_labels: List,
-                                     cutoff: float,
-                                     z_table: AtomicNumberTable, 
-                                     system_selection: str = None,
-                                     environment_selection: str = None,
-                                     subsystem_selection: str = None,
-                                     load_args : dict = None,
-                                     lengths_conversion : float = 10,
-                                     buffer: float = 0.0,
-                                     long_range_cutoff: float = -1.0,
-                                     atom_names: List = None,
-                                     remove_isolated_nodes: bool = False,
-                                     show_progress: bool = True,
-                                     ):
-    # create configurations objects from trajectories
-    configurations = []
-    for i in range(len(trajectories)):
-            configuration = _configurations_from_trajectory(
-                trajectory=trajectories[i],
-                graph_labels=graph_labels[i],
-                node_labels=node_labels[i],
-                system_selection=system_selection,
-                environment_selection=environment_selection,
-                subsystem_selection=subsystem_selection,
-                start=load_args[i]['start'] if load_args is not None else 0,
-                stop=load_args[i]['stop']  if load_args is not None else None,
-                stride=load_args[i]['stride']  if load_args is not None else 1,
-                lengths_conversion=lengths_conversion,
-            )
-            configurations.extend(configuration)
-
-    # convert configurations into DictDataset
-    dataset = create_dataset_from_configurations(
-        config=configurations,
-        z_table=z_table,
-        cutoff=cutoff,
-        buffer=buffer,
-        long_range_cutoff=long_range_cutoff,
-        atom_names=atom_names,
-        remove_isolated_nodes=remove_isolated_nodes,
-        show_progress=show_progress
-    )
-    return dataset
-
-
-def _names_from_top(top: List[mdtraj.Topology] ):
-    it = iter(top)
-    atom_names = list(next(it).atoms)
-    if not all([atom_names == list(n.atoms) for n in it]):
-        raise ValueError(
-            "The atoms names or their order are different in the topology files. Check or deactivate save_names"
-        )
-    
-    return atom_names
-
-
-def _z_table_from_top(
-    top: List[mdtraj.Topology]
-) -> AtomicNumberTable:
-    """
-    Create an atomic number table from the topologies.
-
-    Parameters
-    ----------
-    top: List[mdtraj.Topology]
-        The topology objects.
-    """
-    atomic_numbers = []
-    for t in top:
-        atomic_numbers.extend([a.element.number for a in t.atoms])
-    # atomic_numbers = np.array(atomic_numbers, dtype=int)
-    z_table = AtomicNumberTable.from_zs(atomic_numbers)
-    return z_table
-
-
-def _configurations_from_trajectory(
-    trajectory: mdtraj.Trajectory,
-    graph_labels = None,
-    node_labels = None,
-    system_selection: str = None,
-    environment_selection: str = None,
-    subsystem_selection: str = None,
-    start: int = 0,
-    stop: int = None,
-    stride: int = 1,
-    lengths_conversion : float = 10.0) -> Configurations:
-    """
-    Create configurations from one trajectory.
-
-    Parameters
-    ----------
-    trajectory: mdtraj.Trajectory
-        The MDTraj Trajectory object.
-    graph_labels: np.ndarray
-        Frame-level graph labels for selected frames of this trajectory.
-    system_selection: str
-        MDTraj style atom selections of the system atoms. If given, only
-        selected atoms will be loaded from the trajectories. This option may
-        increase the speed of building graphs.
-    environment_selection: str
-        MDTraj style atom selections of the environment atoms. If given,
-        only the system atoms and [the environment atoms within the cutoff
-        radius of the system atoms] will be kept in the graph.
-    subsystem_selection: str
-        MDTraj style atom selections of the subsystem atoms. If given, long
-        edges will be put between subsystem atoms.
-    lengths_conversion: float,
-        Conversion factor for length units, by default 10.
-        MDTraj uses nanometers, the default sends to Angstroms.
-    """
-    if system_selection is not None and environment_selection is not None:
-        system_atoms = trajectory.top.select(system_selection)
-        assert len(system_atoms) > 0, (
-            'No atoms will be selected with `system_selection`: '
-            + '"{:s}"!'.format(system_selection)
-        )
-        environment_atoms = trajectory.top.select(environment_selection)
-        assert len(environment_atoms) > 0, (
-            'No atoms will be selected with `environment_selection`: '
-            + '"{:s}"!'.format(environment_selection)
-        )
-    else:
-        system_atoms = None
-        environment_atoms = None
-
-    if subsystem_selection is not None:
-        subsystem_atoms = trajectory.top.select(subsystem_selection)
-        assert len(subsystem_atoms) > 0, (
-            'No atoms will be selected with `subsystem_selection`: '
-            + '"{:s}"!'.format(subsystem_selection)
-        )
-        # NOTE: in the above step we have done the atom_slice, so if no
-        # environment atom has been defined, the subsystem atoms will
-        # have to be selected from the sliced atoms, which are previously
-        # defined by system_selection.
-        # So here we only check if the subsystem_selection contains environment
-        # atoms, under the case where both system_selection AND
-        # environment_selection have been given.
-        if system_selection is not None and environment_selection is not None:
-            assert set(subsystem_atoms).issubset(set(system_atoms)), (
-                "All atoms selected by `subsystem_selection` should also be "
-                + "selected by `system_selection`!"
-            )
-    else:
-        subsystem_atoms = None
-
-    atomic_numbers = [a.element.number for a in trajectory.top.atoms]
-    if trajectory.unitcell_vectors is not None:
-        pbc = [True] * 3
-        cell = trajectory.unitcell_vectors
-    else:
-        pbc = [False] * 3
-        cell = [None] * len(trajectory)
-
-    if stop is None:
-        stop = len(trajectory)
-
-    configurations = []
-    frame_indices = list(range(start, stop, stride))
-
-    for local_idx, i in enumerate(frame_indices):
-        label_i = None
-        if graph_labels is not None:
-            label_i = _to_torch_tensor(graph_labels[local_idx]).reshape(-1, 1)
-
-        node_i = None
-        if node_labels is not None:
-            node_i = _to_torch_tensor(node_labels[local_idx])
-            if node_i.ndim == 1:
-                node_i = node_i.reshape(-1, 1)
-
-        configuration = Configuration(
-            atomic_numbers=atomic_numbers,
-            positions=trajectory.xyz[i] * lengths_conversion,
-            cell=cell[i] * lengths_conversion,
-            pbc=pbc,
-            graph_labels=label_i,
-            node_labels=node_i,
-            system=system_atoms,
-            environment=environment_atoms,
-            subsystem=subsystem_atoms,
-        )
-        configurations.append(configuration)
-
-    return configurations
-
 
 
 def test_datasesetFromTrajectories():
